refactor #29: 패키지 명확하게 구분
