+++ src/main/java/ei/algobaroapi/domain/auth/dto/request/AuthSignInRequest.java
@@ -1,4 +1,4 @@
-package ei.algobaroapi.domain.auth.dto;
+package ei.algobaroapi.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
