+++ src/main/java/ei/algobaroapi/domain/auth/dto/response/AuthSignInResponse.java
@@ -1,4 +1,4 @@
-package ei.algobaroapi.domain.auth.dto;
+package ei.algobaroapi.domain.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
